Update the link to the AddressFinder blog
Unfortunately, it appears that Medium (the host of our blog) have changed the format of the URLs for individual blog posts. This PR updates the link to correct this issue. We also took the opportunity to use `https` in the URL, and to also refer to the "dot io" domain we've recently started using.